- fix two compile warnings
git-svn-id: http://svn.ratbox.org/svnroot/trunk@7153 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/m_eob.c
@@ -65,7 +65,7 @@ static void ms_eob(struct Client *client_p, struct Client *source_p,
 {
    sendto_realops_flags(
        FLAGS_ALL,"*** End of burst from %s (%d seconds)",
-       source_p->name, CurrentTime - source_p->firsttime);
+       source_p->name, (signed int)(CurrentTime - source_p->firsttime));
 
    SetEob(client_p);
 }
